blog字段更新

+++ wdoto/blog.py
@@ -1,23 +1,33 @@
-from modleBean.models import Blog, BlogView, BlogFavorite
+from django.utils.datastructures import MultiValueDictKeyError
+
+from modleBean.models import Blog, BlogView, BlogFavorite, BlogRemark, AccountImage
 from django.http import HttpResponse
 from django.forms.models import model_to_dict
 import json
 from rest_framework.permissions import AllowAny, IsAuthenticatedOrReadOnly, IsAuthenticated
 from rest_framework.decorators import api_view, permission_classes
+from django.contrib.auth.models import User
 
 
-# post
+@api_view(['POST'])
+@permission_classes((IsAuthenticated,))
 def blogAdd(request):
     resp = {'code': 0, 'msg': '0'}
     try:
         title = request.POST["title"]
-        ltitle = request.POST["ltitle"]
         body = request.POST["body"]
-        author = request.POST["author"]
-        pageViews = request.POST["pageViews"]
-        favorite = request.POST["favorite"]
-        blog = Blog(title=title, ltitle=ltitle, body=body, author=author, pageViews=pageViews, favorite=favorite)
+        authorId = request.user.id
+        pageViews = request.POST.get("pageViews")
+        favorite = request.POST.get("favorite")
+        if pageViews == None:
+            pageViews = 0
+        if favorite == None:
+            favorite = 0
+        blog = Blog(title=title, body=body, authorId=authorId, pageViews=pageViews, favorite=favorite)
         blog.save()
+    except MultiValueDictKeyError:
+        resp['code'] = 1
+        resp['msg'] = '参数缺失'
     except:
         resp['code'] = 1
     finally:
@@ -30,8 +40,20 @@ def blogDelete(request):
     resp = {'code': 0, 'msg': '0'}
     try:
         id = request.POST['id']
+        authorId = request.user.id
         blog = Blog.objects.get(id=id)
-        blog.delete()
+        if blog == None:
+            resp['code'] = 2
+            resp['msg'] = '找不到id'
+
+        elif blog.authorId != authorId:
+            resp['code'] =   3
+            resp['msg'] = "只有作者能删除"
+        else:
+            blog.delete()
+    except MultiValueDictKeyError:
+        resp['code'] = 1
+        resp['msg'] = '参数缺失'
     except:
         resp['code'] = 1
     finally:
@@ -41,22 +63,31 @@ def blogDelete(request):
 def blogUpdate(request):
     pass
 
+
 def blogList(request):
     print("blogList")
     resp = {'code': 0, 'msg': '0'}
     try:
-        blogs = Blog.objects.all()
+        blogs = Blog.objects.all().order_by("-id")
         print(blogs)
         bl = []
         for blog in blogs:
-            bl.append(model_to_dict(blog))
+            blogRes = model_to_dict(blog)
+            blogRes['date'] = str(blog.date.strftime("%m-%d %H:%I"))
+            user = User.objects.get(id=blog.authorId)
+            blogRes['authorName'] = user.username
+            AccImg = AccountImage.objects.filter(userId=blog.authorId).order_by("-id")
+            blogRes['authorImg'] = str(AccImg[0].image)
+            bl.append(blogRes)
         resp['content'] = bl
         print(resp)
+    except MultiValueDictKeyError:
+        resp['code'] = 1
+        resp['msg'] = '参数缺失'
     except:
         resp['code'] = 1
     finally:
         return HttpResponse(json.dumps(resp), content_type="application/json")
-
 
 
 def blogDetails(request):
@@ -65,7 +96,9 @@ def blogDetails(request):
         id = request.GET['id']
         blog = Blog.objects.get(id=id)
         resp['content'] = model_to_dict(blog)
-
+    except MultiValueDictKeyError:
+        resp['code'] = 1
+        resp['msg'] = '参数缺失'
     except:
         resp['code'] = 1
     finally:
@@ -83,6 +116,9 @@ def blogPageView(request):
         blog.save()
         blogView = BlogView(userId=request.user.id, blogId=id)
         blogView.save()
+    except MultiValueDictKeyError:
+        resp['code'] = 1
+        resp['msg'] = '参数缺失'
     except:
         resp['code'] = 1
     finally:
@@ -100,7 +136,53 @@ def blogFavorite(request):
         blog.save()
         blogFav = BlogFavorite(userId=request.user.id, blogId=id)
         blogFav.save()
+    except MultiValueDictKeyError:
+        resp['code'] = 1
+        resp['msg'] = '参数缺失'
     except:
         resp['code'] = 1
+    finally:
+        return HttpResponse(json.dumps(resp), content_type="application/json")
+
+
+# 权限
+@api_view(['POST'])
+@permission_classes((IsAuthenticated,))
+def blogRemark(request):
+    resp = {'code': 0, 'msg': '0'}
+    try:
+        blogId = request.POST['blogId']
+        userId = request.user.id
+        remark = request.POST['remark']
+        remarkId = request.POST.get("remarkId")
+        blogRemark = BlogRemark(userId=userId, blogId=blogId, remark=remark, remarkId=remarkId)
+        blogRemark.save()
+    except MultiValueDictKeyError:
+        resp['code'] = 1
+        resp['msg'] = '参数缺失'
+    except:
+        resp['code'] = 1
+        resp['msg'] = '其他错误'
+    finally:
+        return HttpResponse(json.dumps(resp), content_type="application/json")
+
+
+# 权限还没加
+@api_view(['POST'])
+@permission_classes((IsAuthenticated,))
+def blogRemarkDel(request):
+    resp = {'code': 0, 'msg': '0'}
+    try:
+        blogId = request.POST['blogId']
+        userId = request.user.id
+        remarkId = request.POST.get("remarkId")
+        blogRemark = BlogRemark.objects.get(userId=userId, id=remarkId, blogId=blogId)
+        blogRemark.save()
+    except MultiValueDictKeyError:
+        resp['code'] = 1
+        resp['msg'] = '参数缺失'
+    except:
+        resp['code'] = 1
+        resp['msg'] = '其他错误'
     finally:
         return HttpResponse(json.dumps(resp), content_type="application/json")
